[FE] refactor: rendering 분리
